Refactor out Deck class
