+++ firefox_puppeteer/firefox_puppeteer/ui/pageinfo/deck.py
@@ -4,28 +4,11 @@
 
 from marionette_driver import By, Wait
 
-from firefox_puppeteer.ui_base_lib import UIBaseLib
+from firefox_puppeteer.ui.deck import BaseDeck
 from firefox_puppeteer.ui.deck import BasePanel
 
 
-class Deck(UIBaseLib):
-
-    def _create_panel_for_id(self, panel_id):
-        """Creates an instance of :class:`Panel` for the specified panel id.
-
-        :param panel_id: The ID of the panel to create an instance of.
-
-        :returns: :class:`Panel` instance
-        """
-        mapping = {'feedPanel': FeedPanel,
-                   'generalPanel': GeneralPanel,
-                   'mediaPanel': MediaPanel,
-                   'permPanel': PermissionsPanel,
-                   'securityPanel': SecurityPanel
-                   }
-
-        panel = self.element.find_element(By.ID, panel_id)
-        return mapping.get(panel_id, Panel)(lambda: self.marionette, self.window, panel)
+class Deck(BaseDeck):
 
     # Properties for visual elements of the deck #
